feat: example auth route
