+++ main.ts
@@ -1,191 +1,70 @@
 import {} from "$ts-rest/open-api"
-import { initContract } from "$ts-rest/core"
-import { z } from "$zod/mod.ts"
+import { createHonoEndpoints, initServer } from "$ts-rest-hono/mod.ts"
+import { sign } from "$hono/middleware/jwt/index.ts"
 
-const c = initContract()
+const seconds = 1000
+const minutes = 60 * seconds
+const hours = 60 * minutes
+const expireAge = 1 * hours
 
-export const userSchema = z.object({
-	jwt: z.string(),
-})
+import env from "./.env.json" with { type: "json" }
+import { Hono } from "$hono/mod.ts"
+import { swaggerUI } from "$hono-swagger-ui"
+import { doc } from "./openapi.ts"
+import { authContract } from "./contract.ts"
 
-export const authContract = c.router({
-	signUp: {
-		method: "POST",
-		path: "/signup",
-		body: z.object({
-			username: z.string(),
-			password: z.string(),
-		}),
-		responses: { 201: null },
-	},
-	signIn: {
-		method: "POST",
-		path: "/signin",
-		body: z.object({
-			username: z.string(),
-			password: z.string(),
-		}),
-		responses: { 200: userSchema },
-	},
-})
-
-const baseHeaders = {
-	baseHeaders: z.object({ Authorization: z.string() }),
+export type JwtPayload = {
+	iat: number
+	sub: string
+	exp: number
+	userId: number
+	name: string
 }
-export const teamContract = c.router({
-	create: {
-		method: "POST",
-		path: "/",
-		body: z.object({
-			name: z.string(),
-		}),
-		responses: { 201: null },
+
+export const authenticate = async (user: {
+	name: string
+	password: string
+	id: number
+}) => {
+	const payload: JwtPayload = {
+		iat: Date.now() / 1000, // 1000 으로 나눠서 초 단위로 만들어줌
+		sub: "wanted social-feed user",
+		exp: Date.now() / 1000 + expireAge,
+		userId: user.id,
+		name: user.name,
+	}
+	const jwt = await sign(payload, env.secret)
+	return jwt
+}
+
+const s = initServer()
+export const authRouter = s.router(authContract, {
+	signUp: async ({ body }) => {
+		// if user exists, return error
+
+		// create new user
+		const user = { id: 1, ...body }
+
+		// return JWT
+
+		const jwt = await authenticate(user)
+		return { status: 201, body: { jwt } }
 	},
-	invite: {
-		method: "POST",
-		path: "/invite",
-		body: z.object({
-			name: z.string(),
-		}),
-		responses: { 201: null },
+	signIn: async ({ body }) => {
+		// find user from DB
+		const user = { id: 1, ...body }
+
+		const jwt = await authenticate(user)
+		return { status: 200, body: { jwt } }
 	},
-	join: {
-		method: "POST",
-		path: "/join",
-		body: z.object({
-			name: z.string(),
-		}),
-		responses: { 201: null },
-	},
-}, {
-	...baseHeaders,
-	pathPrefix: "/team",
 })
 
-export const positiveSchema = z.number().positive()
-export const naturalSchema = positiveSchema.int()
-export const idSchema = z.string().uuid()
-export const tagSchema = z.enum(["Frontend", "Backend", "Design", "QA", "PM", "Document"])
-const ticketSchema = z.object({
-	title: z.string(),
-	position: naturalSchema,
-	tag: tagSchema,
-	hours: positiveSchema,
-	until: z.date(),
-	assignee: idSchema.optional(),
-})
+const app = new Hono()
 
-const columnSchema = z.object({
-	title: z.string(),
-	tickets: z.array(ticketSchema),
-	position: naturalSchema,
-})
+app
+	.get("/openapi.json", (c) => c.json(doc))
+	.get("/openapi", swaggerUI({ url: "/openapi.json" }))
 
-export const ticketContract = c.router({
-	add: {
-		method: "POST",
-		path: "/",
-		body: ticketSchema.omit({ position: true }),
-		responses: { 201: ticketSchema },
-	},
-	reorder: {
-		method: "PUT",
-		path: "/{column}",
-		query: z.object({ to: naturalSchema }),
-		body: null,
-		responses: { 200: null },
-	},
-	remove: {
-		method: "DELETE",
-		path: "/{column}",
-		body: null,
-		responses: { 200: null },
-	},
-	edit: {
-		method: "PATCH",
-		path: "/{column}",
-		body: ticketSchema.omit({ position: true }),
-		responses: { 200: null },
-	},
-}, {
-	...baseHeaders,
-	pathPrefix: "/{board}/{column}",
-})
+createHonoEndpoints(authContract, authRouter, app)
 
-export const columnsContract = c.router({
-	getAll: {
-		method: "GET",
-		path: "/",
-		responses: {
-			200: z.array(columnSchema),
-		},
-	},
-	add: {
-		method: "POST",
-		path: "/",
-		body: columnSchema,
-		responses: { 201: null },
-	},
-	remove: {
-		method: "DELETE",
-		path: "/{column}",
-		body: null,
-		responses: { 200: null },
-	},
-	reorder: {
-		method: "PUT",
-		path: "/{column}",
-		body: z.object({ to: naturalSchema }),
-		responses: { 200: null },
-	},
-	edit: {
-		method: "PATCH",
-		path: "/{column}",
-		body: columnSchema.omit({ position: true }),
-		responses: { 200: null },
-	},
-}, {
-	...baseHeaders,
-	pathPrefix: "/{board}",
-})
-
-export const columnsSchema = z.string()
-	.regex(/^\d+(,\d+)*$/, `"1" 또는 "1,2,3" 형태로 입력해주세요.`)
-	.transform((v) => v.split(",").map(Number))
-
-export const statQuerySchema = z.object({
-	columns: columnsSchema,
-	begin: z.date(),
-	end: z.date(),
-}).partial()
-
-export const statEntrySchema = z.object({
-	id: idSchema,
-	name: z.string(),
-	totalHours: positiveSchema.describe("작업시간 합"),
-	contribution: z.number().describe("기여도"),
-})
-
-export const statsContarct = c.router({
-	oneMember: {
-		method: "GET",
-		path: "/{user}",
-		query: statQuerySchema,
-		responses: {
-			200: z.array(statEntrySchema),
-		},
-	},
-	allMembers: {
-		method: "GET",
-		path: "/",
-		query: statQuerySchema.extend({
-			tag: tagSchema,
-		}),
-		responses: {
-			200: z.array(statEntrySchema),
-		},
-	},
-}, {
-	...baseHeaders,
-	pathPrefix: "/{board}/stats",
-})
+Deno.serve(app.fetch)
